+++ cst_printf.h
@@ -11,7 +11,7 @@ int cst_formats(const char *str, va_list ap, int i);
 int cst_putnbr(int n);
 int cst_print_percent(void);
 int cst_putnbr_u(unsigned long n);
-int cst_print_hex(unsigned long long n, char format);
-int cst_print_addr(unsigned long long n);
+int cst_print_hex(unsigned long n, char format);
+int cst_print_addr(unsigned long n);
 
 #endif
